Parameter Make_with_hash_consing.P

include Intf.Hashed

type t

val compare : t -> t -> int

compare is a total order.

val equal : t -> t -> bool

equal s1 s2 tests whether s1 and s2 are equal.

val hash : t -> int

hash s is a hash of s.

val pp : Stdlib.Format.formatter -> t -> unit

pp fmt s prints a representation of s to the formatter fmt.

val arity : t -> int

arity s is the arity of s, i.e. a term with head s must have exactly arity s sub-terms.